update file structurue
